fix(ci): first-party line limit — split oversized verdict emphasis and test part 026
